\name{fitConic-package}
\alias{fitConic-package}
\docType{package}
\title{
\packageTitle{fitConic}
}
\description{
\packageDescription{fitConic}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fitConic}
\packageIndices{fitConic} 
The main function is \code{fitConic} . 
}
\author{
\packageAuthor{fitConic}, based on code provided in the references and in \code{\link[conicfit:fit.conicLMA]{conicfit::fit.conicLMA()} } 

Maintainer: \packageMaintainer{fitConic}
}
\references{
\url{http://www.mathworks.com/matlabcentral/answers/80541} for the RANSAC-style search to fit rotated parabolas. 
\url{https://math.stackexchange.com/questions/426150} for detailed ellipse parametric equations. 
\url{ https://math.stackexchange.com/questions/2800817} for "focus/directrix/eccentricity" information
\url{https://people.cas.uab.edu/~mosya/cl} and the folks referred to there,
for fitConicLMA .
\url{https://en.wikipedia.org/wiki/Ellipse} for several parameter conversion formulas

A. W. Fitzgibbon, M. Pilu, R. B. Fisher, "Direct Least Squares Fitting of Ellipses", IEEE Trans. PAMI, Vol. 21, pages 476-480 (1999)

Halir R, Flusser J (1998) Proceedings of the 6th International Conference in Central Europe on Computer Graphics and Visualization,  Numerically stable direct least squares fitting of ellipses (WSCG, Plzen, Czech Republic), pp 125132.

}
